--- reducers/index.ts
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux'
-import productReducer from './products'
-import productCarouselReducer from './productCarousels'
-
-const rootReducer = combineReducers({
-    products: productReducer,
-    carousels: productCarouselReducer
-})
-
-export default rootReducer
