+++ src/api/carouselProducts.ts
@@ -1,5 +1,5 @@
-import { carousels } from '../data/carousal'
 import * as _ from 'lodash'
+import { carousels } from '../../data/carousal'
 
 export const getProducts = async () => {
     return Promise.resolve().then(() => {
