+++ src/components/ProductGrid/ProductItem.styles.ts
@@ -1,0 +1,24 @@
+
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+    productContainer: {
+        width: '48%',
+        height: 220,
+        margin: '1%'
+    },
+    thumbnail: {
+        width: '100%',
+        height: 180
+    },
+    title: {
+        color: 'black',
+        alignSelf: 'center'
+    },
+    offer: {
+        color: 'green',
+        alignSelf: 'center'
+    }
+})
+
+export default styles
