+++ src/screens/AllProducts/AllProducts.styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+    screen: {
+        width: '100%',
+        height: '100%'
+    }
+})
+
+export default styles
