+++ src/screens/AllProducts/AllProducts.tsx
@@ -1,26 +1,27 @@
 import React, { useEffect } from 'react'
-import { View, Text, StyleSheet } from 'react-native'
+import { View, ActivityIndicator } from 'react-native'
 import { useSelector, useDispatch } from 'react-redux'
 import ProductGrid from '../../components/ProductGrid/ProductGrid'
 import { getProducts } from '../../actions/products'
+import { AppState } from '../../reducers'
+import styles from './AllProducts.styles'
 
+type ProductState = {
+    isLoading: Boolean
+    products: any[]
+}
 
-const styles = StyleSheet.create({
-    screen: {
-        width: '100%',
-        height: '100%'
+const productSelector = (state: AppState): ProductState => {
+    return {
+        isLoading: state?.products?.isLoading,
+        products: state?.products?.allProducts
     }
-})
+}
 
 
 const AllProducts = () => {
     const dispatch = useDispatch()
-    const { products } = useSelector((state: any) => {
-        return {
-            isLoadingProducts: state['products']['isLoading'],
-            products: state['products']['allProducts']
-        }
-    })
+    const { products, isLoading } = useSelector(productSelector)
 
     useEffect(() => {
         dispatch(getProducts())
@@ -28,6 +29,12 @@ const AllProducts = () => {
 
     const onProductClick = (product) => {
         console.log(product)
+    }
+
+    if (isLoading) {
+        return <View>
+            <ActivityIndicator />
+        </View>
     }
 
     return (
