+++ src/screens/Dashboard/Dashboard.styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+    screen: {
+        width: '100%',
+        height: '100%'
+    },
+    carouselContainer: {
+        width: '100%',
+        height: '25%'
+    }
+})
+
+export default styles
